==> 01_load.qmd <==
---
title: "01_load"
format: html
---

# Load Data for Analysis

**Load Provided Functions**

```{r}
load(file='/home/projects/22140/exam_data/exam_functions.Rdata')
```

**Load Isoform Count Matrix**

```{r}
isoCm <- loadIsoformCountMatrix("GSE132972")
```

```{r}
samples_ocular <- c("GSM3897649","GSM3897650","GSM3897651","GSM3897652","GSM3897653","GSM3897654")
samples_spinal <- c("GSM3897655", "GSM3897656", "GSM3897657","GSM3897658","GSM3897659","GSM3897660")
```

```{r}
isoCm_Ocular <- isoCm[, samples_ocular, drop = FALSE]
isoCm_Spinal <- isoCm[, samples_spinal, drop = FALSE]
```

**Obtain Gene Count Matrix**

```{r}
geneCm_Ocular <- extractGeneCountMatrixFromIsoCounts(isoCm_Ocular)
geneCm_Ocular <- geneCm_Ocular  |> 
  rownames_to_column(var = "gene_id")

geneCm_Spinal <- extractGeneCountMatrixFromIsoCounts(isoCm_Spinal)
geneCm_Spinal <- geneCm_Spinal |> 
  rownames_to_column(var = "gene_id")
```

```{r}
isoInfo <- read_rds(
  paste0(
  '/home/projects/22140/exam_data/',
  'Homo_sapiens.GRCh38.107.chr_patch_hapl_scaff_tx_info.Rds'
  )
)

isoInfo <- isoInfo[ match(rownames(isoCm_Ocular), isoInfo$transcript_id), ]

```

**Get Sample Metadata**

```{r}
sampleMd_ocular <- data.frame(
  sample_id=c("GSM3897649","GSM3897650","GSM3897651","GSM3897652","GSM3897653","GSM3897654"),
  tissue=c("Ocular Motor Neuron", "Ocular Motor Neuron","Ocular Motor Neuron","Ocular Motor Neuron","Ocular Motor Neuron","Ocular Motor Neuron"),
  mutation=c("SOD1","SOD1","SOD1","C9ORF72","C9ORF72","C9ORF72")
  )


sampleMd_spinal <- data.frame(
  sample_id=c("GSM3897655", "GSM3897656", "GSM3897657","GSM3897658","GSM3897659","GSM3897660"),
  tissue=c("Spinal Motor Neuron","Spinal Motor Neuron","Spinal Motor Neuron","Spinal Motor Neuron","Spinal Motor Neuron","Spinal Motor Neuron"),
  mutation=c("SOD1","SOD1","SOD1","C9ORF72","C9ORF72","C9ORF72")
  )
```

## Save Data

```{r}
isoCdf_ocular <- data.frame(isoCm_Ocular)
isoCdf_ocular <- isoCdf_ocular %>%
  rownames_to_column(var = "transcript_id")

isoCdf_spinal <- data.frame(isoCm_Spinal)
isoCdf_spinal <- isoCdf_spinal %>%
  rownames_to_column(var = "transcript_id")
```

```{r}
write_csv(isoCdf_ocular, "data/OcularIsoformCountMatrix.csv")
write_csv(isoCdf_spinal, "data/SpinalIsoformCountMatrix.csv")
write_csv(geneCm_Ocular, "data/OcularGeneCountMatrix.csv")
write_csv(geneCm_Spinal, "data/SpinalGeneCountMatrix.csv")
write_csv(isoInfo, "data/isoformInfo.csv")
write_csv(sampleMd_ocular, "data/OcularSampleMetadata.csv")
write_csv(sampleMd_spinal, "data/SpinalSampleMetadata.csv")
```
